chore: Refactored cli
