+++ apps/cli/package.json
@@ -8,7 +8,7 @@
   "maintainers": [
     "Max Rose (https://www.max-rose.com/)"
   ],
-  "version": "1.0.4",
+  "version": "1.1.0",
   "exports": "./dist/index.js",
   "bin": {
     "envtree": "./dist/index.js"
