+++ apps/cli/src/utils/fileType.ts
@@ -1,0 +1,1 @@
+export type FileType = 'env' | 'json';
