+++ apps/cli/src/utils/typedThis.ts
@@ -1,0 +1,3 @@
+import { Command } from 'commander';
+
+export type TypeThis = Pick<Command, 'processedArgs' | 'opts'>;
