+++ apps/cli/tsconfig.json
@@ -1,5 +1,11 @@
 {
   "extends": "tsconfig/base.json",
+  "compilerOptions": {
+    "baseUrl": ".",
+    "paths": {
+      "@/*": ["src/*"]
+    }
+  },
   "include": [
     "**/*.ts",
     "**/*.tsx"
